// src/api/openspeech/types.ts
export enum MessageType {
  fullClientRequest = 0b0001,
  audioOnlyRequest = 0b0010,
  fullServerResponse = 0b1001,
  audioOnlyServer = 0b1011,
  frontEndResultServer = 0b1100,
  errorResponse = 0b1111,
}

export enum MessageFlagType {
  noSequence = 0b0000,
  positiveSequence = 0b0001,
  lastNoSequence = 0b0010,
  negativeSequence = 0b0011,
  withEvent = 0b0100,
}

export enum SerializationType {
  none = 0b0000,
  json = 0b0001,
}

export enum CompressionType {
  none = 0b0000,
  gzip = 0b0001,
}

export enum ErrorType {
  success = 20000000,
  invalidRequest = 45000001,
  emptyAudio = 45000002,
  timeout = 45000081,
  invalidAudioFormat = 45000151,
  internalError = 55000000,
  busy = 55000031,
}

export enum TtsEventType {
  none = 0,
  startConnection = 1,
  connectionStarted = 50,
  connectionFinished = 52,
  startSession = 100,
  finishSession = 102,
  sessionStarted = 150,
  sessionFinished = 152,
  taskRequest = 200,
  ttsSentenceStart = 350,
  ttsSentenceEnd = 351,
  ttsResponse = 352,
}

export enum ResponseType {
  asrResponse = 'asrResponse',
  ttsResponse = 'ttsResponse',
  errorResponse = 'errorResponse',
}